:bug: ensure config values are cast to int for Carbon

+++ config/trwl.php
@@ -30,12 +30,12 @@ return [
         'time_window' => [
             # time windows before and after a journey to get points
             'good_enough' => [
-                'before' => env('GOOD_ENOUGH_POINTS_MIN_BEFORE', 60),
-                'after'  => env('GOOD_ENOUGH_POINTS_MIN_AFTER', 60),
+                'before' => (int)env('GOOD_ENOUGH_POINTS_MIN_BEFORE', 60),
+                'after'  => (int)env('GOOD_ENOUGH_POINTS_MIN_AFTER', 60),
             ],
             'in_time'     => [
-                'before' => env('FULL_POINTS_MIN_BEFORE', 20),
-                'after'  => env('FULL_POINTS_MIN_AFTER', 10),
+                'before' => (int)env('FULL_POINTS_MIN_BEFORE', 20),
+                'after'  => (int)env('FULL_POINTS_MIN_AFTER', 10),
             ],
         ],
         'train'       => [
